In memory state part 3 (#243)
* In memory remote size queries

* Adding more TCP stuff

* Server test

* Started testing state server

* More tests for state server

* Append message types

* Lock/ unlock

* Append-only

* Multi-chunk push

* Adding tests

* More tests

* More progress on testing

* More refactoring

* More work on remote server test

* More server tests

* Failing tests

* Fixing server test

* Fixing up test

* Clarify local/ global clear

* unused var

+++ include/redis/Redis.h
@@ -125,9 +125,9 @@ namespace redis {
 
         uint32_t acquireLock(const std::string &key, int expirySeconds);
 
-        void releaseLock(const std::string &key, long lockId);
+        void releaseLock(const std::string &key, uint32_t lockId);
 
-        void delIfEq(const std::string &key, long value);
+        void delIfEq(const std::string &key, uint32_t value);
 
         bool setnxex(const std::string &key, long value, int expirySeconds);
